Implement signal and wait commands

+++ src/handler/signal_request.rs
@@ -1,47 +1,22 @@
-use std::collections::HashMap;
+use anyhow::{bail, Result};
+use gprocess_proto::gprocess::api::{
+    response::Command, SignalRequest, SignalResponse,
+};
 
-use gprocess_proto::gprocess::api;
-
-use crate::process_manager::ProcessManager;
-use crate::utils::int_to_signal;
+use crate::{process_manager::ProcessManager, utils::int_to_signal};
 
 pub async fn handle(
-    request: &api::SignalRequest,
+    request: &SignalRequest,
     processes: ProcessManager,
-) -> anyhow::Result<api::response::Command> {
-    todo!()
-
-    /*use api::response::Command;
-
-    let process = match processes.get_mut(&request.pid) {
-        Some(process) => process,
-        None => {
-            return api::Response {
-                request_id,
-                command: Some(Command::Error(api::Error {
-                    message: "Process not found".to_string(),
-                })),
-            };
-        }
-    };
+) -> Result<Command> {
+    if !processes.process_exists(request.pid).await {
+        bail!("pid not found: {}", request.pid);
+    }
 
     let rc = nix::sys::signal::kill(
-        nix::unistd::Pid::from_raw(process.child.id() as i32),
+        nix::unistd::Pid::from_raw(request.pid.try_into()?),
         Some(int_to_signal(request.signal)),
     );
 
-    match rc {
-        Ok(_) => api::Response {
-            request_id,
-            command: Some(Command::Signal(api::SignalResponse {})),
-        },
-        Err(e) => api::Response {
-            request_id,
-            command: Some(Command::Error(api::Error {
-                message: format!("Failed to send signal: {}", e),
-            })),
-        },
-    }
-
-     */
+    Ok(Command::Signal(SignalResponse {}))
 }
